+++ packages/marketplace-ui/src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from "./useBounties";
-export * from "./useCapInfo";
 export * from "./useSampledTransactions";
 export * from "./useLinkTwitter";
 export * from "./useLivePrice";