+++ packages/marketplace-ui/src/index.ts
@@ -2,3 +2,4 @@ export * from "./components";
 export * from "./hooks";
 export * from "./utils";
 export * from "./contexts";
+export * from "./DisburseFunds";