+++ packages/react/src/hooks/useCapInfo.ts
@@ -1,14 +1,20 @@
 import { BN } from "@project-serum/anchor";
 import { PublicKey } from "@solana/web3.js";
-import { useMint, useTokenBonding } from "@strata-foundation/react";
+import { useMint } from "./useMint";
+import { useTokenBonding } from "./useTokenBonding";
 import { toNumber } from "@strata-foundation/spl-token-bonding";
 
+/**
+ * Use mint cap information for a token bonding curve to get information like the number of
+ * items remaining
+ */
 export const useCapInfo = (tokenBondingKey: PublicKey | undefined) => {
   const { info: tokenBonding, loading: loadingBonding } =
     useTokenBonding(tokenBondingKey);
   const targetMintAcct = useMint(tokenBonding?.targetMint);
+  const targetMintSupplyU64 = tokenBonding?.ignoreExternalSupplyChanges ? tokenBonding.supplyFromBonding : targetMintAcct?.supply;
   const targetMintSupply =
-    targetMintAcct && toNumber(targetMintAcct.supply, targetMintAcct);
+    targetMintAcct && toNumber(targetMintSupplyU64, targetMintAcct);
   const mintCap: number | undefined =
     tokenBonding &&
     targetMintAcct &&
@@ -22,6 +28,6 @@ export const useCapInfo = (tokenBondingKey: PublicKey | undefined) => {
   return {
     loading: loadingBonding,
     numRemaining,
-    mintCap
-  }
+    mintCap,
+  };
 };